api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
]